# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Ruth Health Products & Services API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type CreateUserInputRole = typeof CreateUserInputRole[keyof typeof CreateUserInputRole];


export const CreateUserInputRole = {
  member: 'member',
  admin: 'admin',
} as const;

export interface CreateUserInput {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role?: CreateUserInputRole;
  /** @nullable */
  referredByCode?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  guest: 'guest',
  member: 'member',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  promoCode: string;
  /** @nullable */
  referredByCode?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  member: 'member',
  admin: 'admin',
} as const;

export interface UserUpdate {
  name?: string;
  role?: UserUpdateRole;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  /** @nullable */
  imageUrl: string | null;
  guestPrice: number;
  memberPrice: number;
  commissionPct: number;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  guestPrice: number;
  memberPrice: number;
  commissionPct: number;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  /** @nullable */
  imageUrl?: string | null;
  guestPrice?: number;
  memberPrice?: number;
  commissionPct?: number;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  /** @nullable */
  imageUrl: string | null;
  guestPrice: number;
  memberPrice: number;
  commissionPct: number;
  createdAt: string;
}

export interface ServiceInput {
  name: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  guestPrice: number;
  memberPrice: number;
  commissionPct: number;
}

export interface ServiceUpdate {
  name?: string;
  description?: string;
  /** @nullable */
  imageUrl?: string | null;
  guestPrice?: number;
  memberPrice?: number;
  commissionPct?: number;
}

export interface Course {
  id: number;
  name: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  contentUrl?: string | null;
  /** @nullable */
  contentBody?: string | null;
  /** @nullable */
  isEnrolled?: boolean | null;
  createdAt: string;
}

export interface CourseInput {
  name: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  contentUrl?: string | null;
  /** @nullable */
  contentBody?: string | null;
}

export interface CourseUpdate {
  name?: string;
  description?: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  contentUrl?: string | null;
  /** @nullable */
  contentBody?: string | null;
}

export interface Testimonial {
  id: number;
  name: string;
  /** @nullable */
  photoUrl?: string | null;
  text: string;
  createdAt: string;
}

export interface TestimonialInput {
  name: string;
  /** @nullable */
  photoUrl?: string | null;
  text: string;
}

export interface TestimonialUpdate {
  name?: string;
  /** @nullable */
  photoUrl?: string | null;
  text?: string;
}

export type OrderInputItemType = typeof OrderInputItemType[keyof typeof OrderInputItemType];


export const OrderInputItemType = {
  product: 'product',
  service: 'service',
} as const;

export interface ShippingAddress {
  line1?: string;
  city?: string;
  province?: string;
  postalCode?: string;
  country: string;
}

export interface OrderInput {
  itemType: OrderInputItemType;
  itemId: number;
  /** @nullable */
  promoCodeUsed?: string | null;
  shippingAddress: ShippingAddress;
}

export type OrderItemType = typeof OrderItemType[keyof typeof OrderItemType];


export const OrderItemType = {
  product: 'product',
  service: 'service',
} as const;

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
  refunded: 'refunded',
} as const;

export interface Order {
  id: number;
  /** @nullable */
  userId?: number | null;
  itemType: OrderItemType;
  itemId: number;
  /** @nullable */
  itemName?: string | null;
  /** @nullable */
  promoCodeUsed?: string | null;
  itemAmount: number;
  shippingFee: number;
  totalAmount: number;
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  stripePaymentId?: string | null;
  status: OrderStatus;
  createdAt: string;
}

export interface CheckoutSession {
  checkoutUrl: string;
  orderId: number;
}

export interface BookingSlot {
  id: number;
  date: string;
  time: string;
  isBooked: boolean;
}

export interface BookingSlotInput {
  date: string;
  time: string;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  booked: 'booked',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  slotId: number;
  /** @nullable */
  slotDate?: string | null;
  /** @nullable */
  slotTime?: string | null;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  zoomLink?: string | null;
  status: BookingStatus;
  createdAt: string;
}

export interface BookingInput {
  slotId: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
}

export type CommissionEventType = typeof CommissionEventType[keyof typeof CommissionEventType];


export const CommissionEventType = {
  sale: 'sale',
  referral: 'referral',
} as const;

export type CommissionEventStatus = typeof CommissionEventStatus[keyof typeof CommissionEventStatus];


export const CommissionEventStatus = {
  pending: 'pending',
  paid: 'paid',
} as const;

export interface CommissionEvent {
  id: number;
  referringMemberId: number;
  /** @nullable */
  referringMemberName?: string | null;
  type: CommissionEventType;
  /** @nullable */
  relatedId?: number | null;
  amount: number;
  status: CommissionEventStatus;
  createdAt: string;
}

export interface ReferrerSummary {
  memberId: number;
  memberName: string;
  promoCode: string;
  totalCommission: number;
  referralCount: number;
}

export interface DashboardSummary {
  totalMembers: number;
  totalOrders: number;
  totalRevenue: number;
  totalCommissionPending: number;
  recentOrders: Order[];
  topReferrers: ReferrerSummary[];
}

export interface MemberDashboard {
  promoCode: string;
  referralLink: string;
  salesCommission: number;
  referralCommission: number;
  totalCommission: number;
  commissionEvents: CommissionEvent[];
  referredMembers: User[];
}

export interface ShippingZone {
  id: number;
  country: string;
  /** @nullable */
  regionOrPostalPrefix?: string | null;
  feeAmount: number;
  isFree: boolean;
}

export interface ShippingZoneInput {
  country: string;
  /** @nullable */
  regionOrPostalPrefix?: string | null;
  feeAmount: number;
  isFree: boolean;
}

export interface ShippingZoneUpdate {
  country?: string;
  /** @nullable */
  regionOrPostalPrefix?: string | null;
  feeAmount?: number;
  isFree?: boolean;
}

export interface ZoomSettings {
  /** @nullable */
  zoomLink: string | null;
}

export interface ZoomSettingsUpdate {
  zoomLink: string;
}