# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ruth Health Products & Services API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # PRODUCTS
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /courses/{id}/enroll:
    post:
      operationId: enrollCourse
      tags: [courses]
      summary: Enroll in a course (requires login)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # TESTIMONIALS
...
  /booking-slots:
    get:
      operationId: listBookingSlots
      tags: [booking-slots]
      summary: List available booking slots
      responses:
        "200":
          description: List of slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingSlot"
    post:
      operationId: createBookingSlot
      tags: [booking-slots]
      summary: Create a booking slot (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingSlotInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingSlot"
  /booking-slots/{id}:
    delete:
      operationId: deleteBookingSlot
      tags: [booking-slots]
      summary: Delete a booking slot (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # BOOKINGS
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings (admin)
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book a slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "409":
          description: Slot already booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}/cancel:
    patch:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # COMMISSION
...
  /commission/{id}/mark-paid:
    patch:
      operationId: markCommissionPaid
      tags: [commission]
      summary: Mark commission as paid (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionEvent"

  # USERS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # AUTH
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, promoCode, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [guest, member, admin]
        promoCode:
          type: string
        referredByCode:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Course:
      type: object
      required: [id, name, description, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        contentUrl:
          type: ["string", "null"]
        contentBody:
          type: ["string", "null"]
        isEnrolled:
          type: ["boolean", "null"]
        createdAt:
          type: string
          format: date-time
    CourseInput:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        contentUrl:
          type: ["string", "null"]
        contentBody:
          type: ["string", "null"]
    CourseUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        contentUrl:
          type: ["string", "null"]
        contentBody:
          type: ["string", "null"]

    # TESTIMONIALS
...
    BookingSlot:
      type: object
      required: [id, date, time, isBooked]
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        time:
          type: string
        isBooked:
          type: boolean
    BookingSlotInput:
      type: object
      required: [date, time]
      properties:
        date:
          type: string
          format: date
        time:
          type: string

    # BOOKINGS
    Booking:
      type: object
      required: [id, slotId, name, email, status, createdAt]
      properties:
        id:
          type: integer
        slotId:
          type: integer
        slotDate:
          type: ["string", "null"]
        slotTime:
          type: ["string", "null"]
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        zoomLink:
          type: ["string", "null"]
        status:
          type: string
          enum: [booked, completed, cancelled]
        createdAt:
          type: string
          format: date-time
    BookingInput:
      type: object
      required: [slotId, name, email]
      properties:
        slotId:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]

    # COMMISSION
    CommissionEvent:
      type: object
      required: [id, referringMemberId, type, amount, status, createdAt]
      properties:
        id:
          type: integer
        referringMemberId:
          type: integer
        referringMemberName:
          type: ["string", "null"]
        type:
          type: string
          enum: [sale, referral]
        relatedId:
          type: ["integer", "null"]
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid]
        createdAt:
          type: string
          format: date-time

    # DASHBOARD
...